refine PP-TSM serving code and add code for TIPC
